Added importValidUsersList
Error checking is still rudimentary so the Users valid logins page needs to be perfect almost. But the hard-coded values have been left as is to preserve core functionality. There is a bug where after a user has been added and written to the file, even though they are added to user list (TBC) we can't login through them.